Add VESSL H100 job entrypoint (scripts/vessl_run.sh) + bf16/dtype support for large models
scripts/vessl_run.sh: GitHub auth via GH_TOKEN secret -> sync branch -> pip deps (slim torch
container, GDN needs no mamba_ssm build) -> GPU+GDN-gate sanity -> run train_grm_passkey.
Parametrized via env (ARCH/VARIANT/MODEL/TRAIN_LEN/DTYPE/...). Targets the H100 (GDN chunk
backward needs >A100 shared memory). Container: quay.io/vessl-ai/torch:2.9.1-cuda12.8.1-py3.10-slim.

Also enables size-scaling on big mamba2: train_grm_passkey --dtype {float32,bfloat16} (heads cast
to match), and base.run_mixer_with_cache casts routing descriptors to the read-out dtype (fp32 scan
states -> bf16 heads) so the read-out einsums don't hit a dtype mismatch.

+++ lmr/adapters/base.py
@@ -111,7 +111,9 @@ def run_mixer_with_cache(adapter: Adapter, mixer, x_in, cached_states, readout, 
         y_i, _ = adapter.scan(mixer, proj, h_i, backend, memory_only=True)
         y_cached.append(y_i)
         descriptors.append(adapter.descriptor(h_i))
-    descriptors = torch.stack(descriptors, dim=1) if descriptors else None
+    # The scan computes states in fp32; cast routing descriptors to the read-out dtype (e.g. bf16
+    # for large models) so the read-out heads' einsums don't hit a dtype mismatch.
+    descriptors = torch.stack(descriptors, dim=1).to(x_in.dtype) if descriptors else None
 
     y, aux = readout(y_main, y_cached, x_in, descriptors)
     out = adapter.finalize(mixer, proj, y, dtype)
